Refactor table into abstract and costTable
