+++ src/components/Package.js
@@ -1,14 +1,14 @@
-import Table from './ux/Table'
+import Table from './CostTable'
 import styles from 'views/skyBill.scss'
 
 export default function Package (props) {
-  const { subscriptions = [], total = 0 } = props
-  const _subscriptions = subscriptions.map(c => ({...c, cost: '£' + c.cost.toFixed(2)}))
+
+  const { subscriptions, total = 0 } = props
 
   return (
     <div>
       <h3>Subscriptions</h3>
-      <Table {..._subscriptions} />
+      <Table {...subscriptions} />
       <p className={styles.total}>Subtotal: &pound;{total.toFixed(2)}</p>
     </div>
   )
