+++ src/components/ux/Table.js
@@ -1,16 +1,19 @@
 import styles from './Table.scss'
+import { toArray } from '../../utils'
 
 export default function Table (data) {
-  data = {}.toString.call(data) === 'Array' ? data : Object.keys(data).map(k => data[k])
-  const keys = data.length ? Object.keys(data[0]) : data
+
+  const values = toArray(data)
+  const keys = values.length ? Object.keys(values[0]) : []
+  const tableClass = styles.table + (data.className? ' ' + data.className : '')
 
   return (
-    <table className={styles.table}>
+    <table className={tableClass}>
       <thead><tr>
       {keys.map(h => <th key={h}>{h}</th>)}
       </tr></thead>
       <tbody>
-        {data.map((b,i) => {
+        {values.map((b,i) => {
           return <tr key={'r'+i}>
             {keys.map(h => <td key={h} >{b[h]}</td>)}
           </tr>
